Test refactoring: use "@JacksonTestFailureExpected" from jackson-core test artifact
