+++ src/test/java/tools/jackson/databind/records/tofix/DuplicatePropertyDeserializationRecord4690Test.java
@@ -2,9 +2,10 @@ package tools.jackson.databind.records.tofix;
 
 import org.junit.jupiter.api.Test;
 
+import tools.jackson.core.testutil.failure.JacksonTestFailureExpected;
+
 import tools.jackson.databind.ObjectMapper;
 import tools.jackson.databind.testutil.DatabindTestUtil;
-import tools.jackson.databind.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -14,7 +15,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 public class DuplicatePropertyDeserializationRecord4690Test
         extends DatabindTestUtil
 {
-
     record MyRecord(String first) { }
 
     private final ObjectMapper mapper = newJsonMapper();
