+++ src/test/java/tools/jackson/databind/tofix/AsEmptyPOJONoDefaultCtor2572Test.java
@@ -3,12 +3,12 @@ package tools.jackson.databind.tofix;
 import java.util.Objects;
 
 import org.junit.jupiter.api.Test;
+import tools.jackson.core.testutil.failure.JacksonTestFailureExpected;
 
 import com.fasterxml.jackson.annotation.*;
 
 import tools.jackson.databind.*;
 import tools.jackson.databind.testutil.DatabindTestUtil;
-import tools.jackson.databind.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
