+++ src/test/java/tools/jackson/databind/tofix/ExternalTypeIdWithUnwrapped2039Test.java
@@ -1,12 +1,12 @@
 package tools.jackson.databind.tofix;
 
 import org.junit.jupiter.api.Test;
+import tools.jackson.core.testutil.failure.JacksonTestFailureExpected;
 
 import com.fasterxml.jackson.annotation.*;
 
 import tools.jackson.databind.*;
 import tools.jackson.databind.testutil.DatabindTestUtil;
-import tools.jackson.databind.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.*;
