+++ src/test/java/tools/jackson/databind/tofix/NoTypeInfo1654Test.java
@@ -3,6 +3,7 @@ package tools.jackson.databind.tofix;
 import java.util.*;
 
 import org.junit.jupiter.api.Test;
+import tools.jackson.core.testutil.failure.JacksonTestFailureExpected;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
@@ -10,7 +11,6 @@ import tools.jackson.core.JsonParser;
 import tools.jackson.databind.*;
 import tools.jackson.databind.annotation.JsonDeserialize;
 import tools.jackson.databind.testutil.DatabindTestUtil;
-import tools.jackson.databind.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
